+++ goplay/src/test/java/be/tapped/goplay/HtmlJsonProgramExtractorTest.kt
@@ -1,7 +1,6 @@
 package be.tapped.goplay
 
 import be.tapped.goplay.content.HtmlJsonProgramExtractor
-import be.tapped.goplay.content.jsonSerializer
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.ShouldSpec
 import io.kotest.matchers.collections.shouldHaveSize
