+++ goplay/src/test/java/be/tapped/goplay/e2e/EpgRepoTest.kt
@@ -1,8 +1,8 @@
 package be.tapped.goplay.e2e
 
 import be.tapped.goplay.GoPlayBrand
-import be.tapped.goplay.content.httpClient
 import be.tapped.goplay.epg.httpEpgRepo
+import be.tapped.goplay.httpClient
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.ShouldSpec
 import kotlinx.datetime.Clock
