+++ goplay/src/test/java/be/tapped/goplay/e2e/ProgramRepoTest.kt
@@ -1,9 +1,9 @@
 package be.tapped.goplay.e2e
 
 import be.tapped.goplay.content.HtmlJsonProgramExtractor
-import be.tapped.goplay.content.httpClient
 import be.tapped.goplay.content.httpProgramRepo
-import be.tapped.goplay.content.jsonSerializer
+import be.tapped.goplay.httpClient
+import be.tapped.goplay.jsonSerializer
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.ShouldSpec
